pyproject.toml
[project]
name = "swuleague"
version = "0.1.0"
description = "Générateur d'images de classement pour ligues Star Wars Unlimited"
requires-python = ">=3.11"
dependencies = [
    "Pillow>=10.0",
]

[project.scripts]
swuleague = "image:main"
